Fixing bad recommendations
