+++ cpp/mrc/include/mrc/modules/sample_modules.hpp
@@ -19,13 +19,16 @@
 
 #include "mrc/core/utils.hpp"
 #include "mrc/modules/segment_modules.hpp"
-#include "mrc/segment/builder.hpp"
 
 #include <nlohmann/json.hpp>
 #include <rxcpp/rx.hpp>
 
 #include <string>
 #include <utility>
+
+namespace mrc::segment {
+class Builder;
+}  // namespace mrc::segment
 
 namespace mrc::modules {
