+++ cpp/mrc/include/mrc/segment/definition.hpp
@@ -18,18 +18,16 @@
 #pragma once
 
 #include "mrc/engine/segment/idefinition.hpp"
-#include "mrc/segment/egress_ports.hpp"
-#include "mrc/segment/ingress_ports.hpp"
+#include "mrc/segment/initializers.hpp"
 #include "mrc/utils/macros.hpp"
 
-#include <functional>
 #include <map>
 #include <memory>
 #include <string>
-#include <utility>
 
 namespace mrc::segment {
-class Builder;
+struct EgressPortsBase;
+struct IngressPortsBase;
 
 class Definition final : public internal::segment::IDefinition
 {
@@ -53,33 +51,17 @@ class Definition final : public internal::segment::IDefinition
     static std::shared_ptr<Definition> create(std::string name,
                                               IngressPortsBase ingress_ports,
                                               EgressPortsBase egress_ports,
-                                              segment_initializer_fn_t initializer)
-    {
-        return Definition::create(
-            std::move(name), ingress_ports.m_initializers, egress_ports.m_initializers, std::move(initializer));
-    }
+                                              segment_initializer_fn_t initializer);
 
     static std::shared_ptr<Definition> create(std::string name,
                                               EgressPortsBase egress_ports,
-                                              segment_initializer_fn_t initializer)
-    {
-        return Definition::create(std::move(name), {}, egress_ports.m_initializers, std::move(initializer));
-    }
+                                              segment_initializer_fn_t initializer);
 
     static std::shared_ptr<Definition> create(std::string name,
                                               IngressPortsBase ingress_ports,
-                                              segment_initializer_fn_t initializer)
-    {
-        return Definition::create(std::move(name), ingress_ports.m_initializers, {}, std::move(initializer));
-    }
+                                              segment_initializer_fn_t initializer);
 
-    static std::shared_ptr<Definition> create(std::string name, segment_initializer_fn_t initializer)
-    {
-        return Definition::create(std::move(name),
-                                  std::map<std::string, ingress_initializer_t>{},
-                                  std::map<std::string, egress_initializer_t>{},
-                                  std::move(initializer));
-    }
+    static std::shared_ptr<Definition> create(std::string name, segment_initializer_fn_t initializer);
 };
 
 }  // namespace mrc::segment
