+++ cpp/mrc/src/public/pipeline/pipeline.cpp
@@ -19,6 +19,7 @@
 
 #include "mrc/segment/egress_ports.hpp"
 #include "mrc/segment/ingress_ports.hpp"
+#include "mrc/segment/definition.hpp"
 
 namespace mrc::pipeline {
